split out worker functionality
